+++ src/test/scala/com/github/thanhtien522/eshttpclient/ApiEntitiesTest.scala
@@ -1,8 +1,8 @@
-package org.elasticsearch.client.http
+package com.github.thanhtien522.eshttpclient
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
-import org.elasticsearch.client.http.entities.SearchResponse
+import com.github.thanhtien522.eshttpclient.entities.SearchResponse
 import org.scalatest.FunSuite
 
 /**
