+++ src/test/scala/com/github/thanhtien522/eshttpclient/EsHttpClientSearchTest.scala
@@ -1,4 +1,4 @@
-package org.elasticsearch.client.http
+package com.github.thanhtien522.eshttpclient
 
 /**
   * Created by user on 6/15/17.
@@ -6,7 +6,8 @@ package org.elasticsearch.client.http
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
-import org.elasticsearch.client.http.entities.{BasicAuthInfo, NoAuth, SearchRequest}
+import com.github.thanhtien522.eshttpclient.entities.SearchRequest
+import org.elasticsearch.client.http.entities.{BasicAuthInfo, NoAuth}
 import org.elasticsearch.index.query.QueryBuilders
 import org.elasticsearch.search.builder.SearchSourceBuilder
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterAllConfigMap, ConfigMap, FunSuite}
